services/telemetry-cli/requirements.txt
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0

